Slit Iso/_conds from Iso/_vars
